Update signature of `stego_gen_keys`

+++ src/stego.c
@@ -26,7 +26,7 @@ static int generate_random_key(uint8_t* buf) {
  * Generating "count" orthonormal keys.
  * Uses special Gram-Schmidt to do so.
  */
-int stego_gen_keys(void* buf, int count) {
+int stego_gen_keys(void* buf, size_t count) {
   size_t total_keys_size = count * STEGO_KEY_SIZE;
   uint8_t* int_buf = (uint8_t*) buf;
   for (size_t i = 0; i < total_keys_size; i += STEGO_KEY_SIZE) {
