+++ src/stego.h
@@ -8,7 +8,7 @@
 #define COVER_FILE_COUNT 128
 #define STEGO_KEY_SIZE (COVER_FILE_COUNT / CHAR_BIT)
 
-int stego_gen_keys(void* buf, int count);
+int stego_gen_keys(void* buf, size_t count);
 
 size_t compute_level_size(size_t disk_size);
